+++ rs/ethereum/cketh/minter/src/eth_logs/parser.rs
@@ -1,0 +1,238 @@
+use crate::eth_logs::{
+    parse_principal_from_slice, EventSource, EventSourceError, LedgerSubaccount,
+    ReceivedErc20Event, ReceivedEthEvent, ReceivedEvent, ReceivedEventError,
+};
+use crate::eth_rpc::{Data, FixedSizeData, LogEntry};
+use crate::numeric::{BlockNumber, Erc20Value, Wei};
+use candid::Principal;
+use ic_ethereum_types::Address;
+
+/// Parse an Ethereum log event into a `ReceivedEvent`.
+pub trait LogParser {
+    fn parse_log(log: LogEntry) -> Result<ReceivedEvent, ReceivedEventError>;
+}
+
+pub struct ReceivedEthLogParser {}
+
+impl LogParser for ReceivedEthLogParser {
+    fn parse_log(entry: LogEntry) -> Result<ReceivedEvent, ReceivedEventError> {
+        let (block_number, event_source) = ensure_not_pending(&entry)?;
+        ensure_not_removed(&entry, event_source)?;
+
+        ensure_topics(
+            &entry,
+            |topics| {
+                topics.len() == 3
+                    && topics.first()
+                        == Some(&FixedSizeData(crate::deposit::RECEIVED_ETH_EVENT_TOPIC))
+            },
+            event_source,
+        )?;
+        let from_address = parse_address(&entry.topics[1], event_source)?;
+        let principal = parse_principal(&entry.topics[2], event_source)?;
+
+        let [value_bytes] = parse_data_into_32_byte_words(entry.data, event_source)?;
+        let EventSource {
+            transaction_hash,
+            log_index,
+        } = event_source;
+
+        Ok(ReceivedEthEvent {
+            transaction_hash,
+            block_number,
+            log_index,
+            from_address,
+            value: Wei::from_be_bytes(value_bytes),
+            principal,
+            subaccount: None,
+        }
+        .into())
+    }
+}
+
+pub struct ReceivedErc20LogParser {}
+
+impl LogParser for ReceivedErc20LogParser {
+    fn parse_log(entry: LogEntry) -> Result<ReceivedEvent, ReceivedEventError> {
+        let (block_number, event_source) = ensure_not_pending(&entry)?;
+        ensure_not_removed(&entry, event_source)?;
+
+        ensure_topics(
+            &entry,
+            |topics| {
+                topics.len() == 4
+                    && topics.first()
+                        == Some(&FixedSizeData(crate::deposit::RECEIVED_ERC20_EVENT_TOPIC))
+            },
+            event_source,
+        )?;
+        let erc20_contract_address = parse_address(&entry.topics[1], event_source)?;
+        let from_address = parse_address(&entry.topics[2], event_source)?;
+        let principal = parse_principal(&entry.topics[3], event_source)?;
+
+        let [value_bytes] = parse_data_into_32_byte_words(entry.data, event_source)?;
+        let EventSource {
+            transaction_hash,
+            log_index,
+        } = event_source;
+
+        Ok(ReceivedErc20Event {
+            transaction_hash,
+            block_number,
+            log_index,
+            from_address,
+            value: Erc20Value::from_be_bytes(value_bytes),
+            principal,
+            erc20_contract_address,
+            subaccount: None,
+        }
+        .into())
+    }
+}
+
+pub struct Erc20WithSubaccountLogParser {}
+
+impl LogParser for Erc20WithSubaccountLogParser {
+    fn parse_log(entry: LogEntry) -> Result<ReceivedEvent, ReceivedEventError> {
+        let (block_number, event_source) = ensure_not_pending(&entry)?;
+        ensure_not_removed(&entry, event_source)?;
+
+        ensure_topics(
+            &entry,
+            |topics| {
+                topics.len() == 4
+                    && topics.first()
+                        == Some(&FixedSizeData(
+                            crate::deposit::RECEIVED_ERC20_EVENT_WITH_SUBACCOUNT_TOPIC,
+                        ))
+            },
+            event_source,
+        )?;
+        let erc20_contract_address = parse_address(&entry.topics[1], event_source)?;
+        let from_address = parse_address(&entry.topics[2], event_source)?;
+        let principal = parse_principal(&entry.topics[3], event_source)?;
+        let [value_bytes, subaccount_bytes] =
+            parse_data_into_32_byte_words(entry.data, event_source)?;
+        let value = Erc20Value::from_be_bytes(value_bytes);
+        let subaccount = LedgerSubaccount::from_bytes(subaccount_bytes);
+        let EventSource {
+            transaction_hash,
+            log_index,
+        } = event_source;
+
+        Ok(ReceivedErc20Event {
+            transaction_hash,
+            block_number,
+            log_index,
+            from_address,
+            value,
+            principal,
+            erc20_contract_address,
+            subaccount,
+        }
+        .into())
+    }
+}
+
+fn ensure_not_pending(entry: &LogEntry) -> Result<(BlockNumber, EventSource), ReceivedEventError> {
+    let _block_hash = entry
+        .block_hash
+        .ok_or(ReceivedEventError::PendingLogEntry)?;
+    let block_number = entry
+        .block_number
+        .ok_or(ReceivedEventError::PendingLogEntry)?;
+    let transaction_hash = entry
+        .transaction_hash
+        .ok_or(ReceivedEventError::PendingLogEntry)?;
+    let _transaction_index = entry
+        .transaction_index
+        .ok_or(ReceivedEventError::PendingLogEntry)?;
+    let log_index = entry.log_index.ok_or(ReceivedEventError::PendingLogEntry)?;
+    Ok((
+        block_number,
+        EventSource {
+            transaction_hash,
+            log_index,
+        },
+    ))
+}
+
+fn ensure_not_removed(
+    entry: &LogEntry,
+    event_source: EventSource,
+) -> Result<(), ReceivedEventError> {
+    if entry.removed {
+        return Err(ReceivedEventError::InvalidEventSource {
+            source: event_source,
+            error: EventSourceError::InvalidEvent(
+                "this event has been removed from the chain".to_string(),
+            ),
+        });
+    }
+    Ok(())
+}
+
+fn ensure_topics<P>(
+    entry: &LogEntry,
+    predicate: P,
+    event_source: EventSource,
+) -> Result<(), ReceivedEventError>
+where
+    P: FnOnce(&[FixedSizeData]) -> bool,
+{
+    if !predicate(&entry.topics) {
+        return Err(ReceivedEventError::InvalidEventSource {
+            source: event_source,
+            error: EventSourceError::InvalidEvent("Invalid topics".to_string()),
+        });
+    }
+    Ok(())
+}
+
+fn parse_address(
+    address: &FixedSizeData,
+    event_source: EventSource,
+) -> Result<Address, ReceivedEventError> {
+    Address::try_from(&address.0).map_err(|err| ReceivedEventError::InvalidEventSource {
+        source: event_source,
+        error: EventSourceError::InvalidEvent(format!("Invalid address in log entry: {}", err)),
+    })
+}
+
+fn parse_principal(
+    principal: &FixedSizeData,
+    event_source: EventSource,
+) -> Result<Principal, ReceivedEventError> {
+    parse_principal_from_slice(&principal.0).map_err(|_err| {
+        ReceivedEventError::InvalidEventSource {
+            source: event_source,
+            error: EventSourceError::InvalidPrincipal {
+                invalid_principal: principal.clone(),
+            },
+        }
+    })
+}
+
+fn parse_data_into_32_byte_words<const N: usize>(
+    data: Data,
+    event_source: EventSource,
+) -> Result<[[u8; 32]; N], ReceivedEventError> {
+    let data = data.0;
+    if data.len() != 32 * N {
+        return Err(ReceivedEventError::InvalidEventSource {
+            source: event_source,
+            error: EventSourceError::InvalidEvent(format!(
+                "Expected {} bytes, got {}",
+                32 * N,
+                data.len()
+            )),
+        });
+    }
+    let mut result = Vec::with_capacity(N);
+    for chunk in data.chunks_exact(32) {
+        let mut word = [0; 32];
+        word.copy_from_slice(chunk);
+        result.push(word);
+    }
+    Ok(result.try_into().unwrap())
+}
